vk adding

+++ app/src/main/java/com/example/finish1m/Domain/Models/WallModels/CopyHistory.java
@@ -18,7 +18,7 @@ public class CopyHistory {
     public int fromId;
     @SerializedName("date")
     @Expose
-    public int date;
+    public long date;
     @SerializedName("post_type")
     @Expose
     public String postType;
