+++ app/src/main/java/com/example/finish1m/Domain/Models/WallModels/Item.java
@@ -18,7 +18,7 @@ public class Item {
     public int ownerId;
     @SerializedName("date")
     @Expose
-    public int date;
+    public long date;
     @SerializedName("marked_as_ads")
     @Expose
     public int markedAsAds;
